Added first coment for doc, ROS and minor fixes

+++ control_window_params.cpp
@@ -12,19 +12,19 @@ void parameterController::printCurrentConfig() {
     std::cout << "_________________________________________" << std::endl;
 }
 
-void parameterController::loadConfig(int configNr) {
+void parameterController::loadConfig(unsigned long configNr) {
     std::cout << "Loading Config " << configNr << " as current Config!" << std::endl;
     std::vector<std::vector<std::string>> configData;
     readInData("param_config.pacf", configData);
     
-    if(configNr < 0 || configNr > 9 || (configData.size() < (configNr+1))) {
+    if(configNr < 0 || configNr > 9 || (configData.size() < (configNr + 1))) {
         std::cerr << "\n\tAn Error occured during loading of the Config!!\n\t Please Try Again!!" << std::endl; 
         return;
     }
 
     std::vector<std::string> singleConfig = configData[configNr];
 
-    for(int i = 1; i < singleConfig.size() - 1; i++) {
+    for(unsigned long i = 1; i < singleConfig.size() - 1; i++) {
         std::vector<std::string> seperatedData = seperateString(singleConfig[i], "#");
         descriptors[i-1]->selectedValue = std::stoi(seperatedData[1]);
     }
